Folder structure refactoring
